근처샵, 샵, 샵상세 완료 이미징 dimen 조금더 내일 timeline추가 맵 작업하면 데이터 수신은 거의 끝남
데이터 송신부분은 물어봐야하곘다.

+++ app/src/main/java/com/tacademy/qodbtn41/gosurf/data/Locations.java
@@ -4,8 +4,8 @@ package com.tacademy.qodbtn41.gosurf.data;
  * Created by UserPC on 2015-11-10.
  */
 public class Locations {
-    int latitude;
-    int longitude;
+    double latitude;
+    double longitude;
 
     public void setLatitude(int latitude) {
         this.latitude = latitude;
@@ -15,11 +15,11 @@ public class Locations {
         this.longitude = longitude;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public int getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 }
